+++ src/main/java/com/ba/controller/WaiterController.java
@@ -1,6 +1,7 @@
 package com.ba.controller;
 
 import com.ba.dto.WaiterDTO;
+import com.ba.exception.BusinessRuleException;
 import com.ba.service.WaiterService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -22,16 +23,25 @@ public class WaiterController {
 
     @PostMapping("/add-waiter")
     public String addWaiter(@RequestBody WaiterDTO waiterDTO){
+        if(waiterDTO == null){
+            throw new BusinessRuleException("Waiter cannot be empty...!");
+        }
         return waiterService.addWaiter(waiterDTO);
     }
 
     @DeleteMapping("/delete-waiter/{id}")
     public String deleteWaiter(@PathVariable Long id){
+        if(id == null){
+            throw new BusinessRuleException("Id cannot be empty...!");
+        }
         return  waiterService.deleteWaiter(id);
     }
 
     @PutMapping("/update-waiter")
     public WaiterDTO updateWaiter(@RequestBody WaiterDTO waiterDTO){
+        if(waiterDTO == null){
+            throw new BusinessRuleException("Waiter cannot be empty...!");
+        }
         return waiterService.updateWaiter(waiterDTO);
     }
 }
